requirements.txt
Flask>=3.0.0
pandas>=2.2.0
plotly>=5.18.0
dash>=2.16.0
dash-bootstrap-components>=1.5.0
python-dateutil>=2.8.0
